Changes to certain a Check
Move Pirate checker to check first thing on load of QModLoader.
Add log line if not detecting pirates too so we can tell if the line was deleted.
Added QModService for checking Pirated so mods can use it to mess with pirates if they so choose.

Added a special dialog for later if we ever decide to change how we work with pirates.